1.修改按钮问题修改 1.修改在提交pdf时调用 1.修改在最后审批人时 监听器调用

+++ ruoyi-modules/ruoyi-modules-danny/src/main/java/com/ruoyi/danny/service/impl/DanWorkflowReimServiceImpl.java
@@ -169,26 +169,30 @@ public class DanWorkflowReimServiceImpl implements IDanWorkflowReimService
     public int updateDanWorkflowReim(DanWorkflowReim danWorkflowReim)
     {
         danWorkflowReim.setUpdateTime(DateUtils.getNowDate());
-        danWorkflowReimMapper.deleteDanWorkflowReimgoodsByReimId(danWorkflowReim.getReimId());
-        //
-        if(danWorkflowReim.getReimId()==null){ //避免ReimEndStateListener的notify调用执行
-            List<DanWorkflowReimgoods> danWorkflowReimgoods = danWorkflowReim.getDanWorkflowReimgoodsList();
-            BigDecimal amount = new BigDecimal(0);
-            BigDecimal balance = new BigDecimal(0);
-            BigDecimal anleihen = new BigDecimal(0);
-            for(DanWorkflowReimgoods reimgood : danWorkflowReimgoods){
-                amount = amount.add(reimgood.getMoney());
-            }
-            if(danWorkflowReim.getAnleihen()==null){
-                danWorkflowReim.setAnleihen(anleihen);
-            }
-            balance =amount.subtract(danWorkflowReim.getAnleihen());
 
-            danWorkflowReim.setAmount(amount);
-            danWorkflowReim.setAnleihen(danWorkflowReim.getAnleihen());
-            danWorkflowReim.setBalance(balance);
+        //只有当表单子项不为null时才更新 避免避免ReimEndStateListener的notify调用执行
+        if(danWorkflowReim.getDanWorkflowReimgoodsList()!=null){
+            danWorkflowReimMapper.deleteDanWorkflowReimgoodsByReimId(danWorkflowReim.getReimId());
+            List<DanWorkflowReimgoods> danWorkflowReimgoods = danWorkflowReim.getDanWorkflowReimgoodsList();
+            if(danWorkflowReim.getReimId()==null){ //避免ReimEndStateListener的notify调用执行
+                BigDecimal amount = new BigDecimal(0);
+                BigDecimal balance = new BigDecimal(0);
+                BigDecimal anleihen = new BigDecimal(0);
+                for(DanWorkflowReimgoods reimgood : danWorkflowReimgoods){
+                    amount = amount.add(reimgood.getMoney());
+                }
+                if(danWorkflowReim.getAnleihen()==null){
+                    danWorkflowReim.setAnleihen(anleihen);
+                }
+                balance =amount.subtract(danWorkflowReim.getAnleihen());
+
+                danWorkflowReim.setAmount(amount);
+                danWorkflowReim.setAnleihen(danWorkflowReim.getAnleihen());
+                danWorkflowReim.setBalance(balance);
+            }
+            insertDanWorkflowReimgoods(danWorkflowReim);
         }
-        insertDanWorkflowReimgoods(danWorkflowReim);
+
 
         /**
          * 给流程实例加入变量 这里附件的地址
